% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_currency_convert.R
\name{col_currency_convert}
\alias{col_currency_convert}
\title{Column Currency Conversion}
\usage{
col_currency_convert(df, col_code, col_value, to = "USD", key)
}
\arguments{
\item{df}{The data frame containing the values to be converted.}

\item{col_code}{The name of the column in the data frame containing the currency codes. Ensure you wrap the name of the column in "".}

\item{key}{Your openexchangerates.org API Key. Ensure you wrap the API key "".}

\item{col_values}{The name of the column in the data frame containing the currency codes. Ensure you wrap the name of the column in "".}
}
\description{
This function converts currencies in a data frame. It requires currency codes to be in a seperate column than the currency values.
}
\examples{
col_currency_convert()
}
\keyword{convert}
\keyword{currency,}
\keyword{lucr,}
